+++ EssentialFeed/EssentialFeedTests/Feed Presentation/FeedPresenterTests.swift
@@ -7,84 +7,6 @@
 import XCTest
 import EssentialFeed
 
-struct FeedViewModel {
-    let feed: [FeedImage]
-}
-
-struct FeedLoadingViewModel {
-    let isLoading: Bool
-}
-
-protocol FeedLoadingView {
-     func display(_ viewModel: FeedLoadingViewModel)
-}
-
-protocol FeedView {
-     func display(_ vieModel: FeedViewModel)
-}
-
-struct FeedErrorViewModel {
-    
-    let message: String?
-    
-    static var noError: FeedErrorViewModel {
-        return FeedErrorViewModel(message: nil)
-    }
-    
-    static func error(message: String) -> FeedErrorViewModel {
-        return FeedErrorViewModel(message: message)
-    }
-}
-
-protocol FeedErrorView {
-    func display(_ viewModel: FeedErrorViewModel)
-}
-
-final class FeedPresenter {
-    private let feedView: FeedView
-    private let errorView: FeedErrorView
-    private let loadingView: FeedLoadingView
-    
-    private var feedLoadError: String {
-        return NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
-            tableName: "Feed",
-            bundle: Bundle(for: FeedPresenter.self),
-            comment: "Error message displayed when we can't load the image feed from the server")
-    }
-
-    init(feedView: FeedView, loadingView: FeedLoadingView, errorView: FeedErrorView) {
-        self.feedView = feedView
-        self.loadingView = loadingView
-        self.errorView = errorView
-    }
-    
-    static var title: String {
-        return NSLocalizedString(
-           "FEED_VIEW_TITLE",
-           tableName: "Feed",
-           bundle: Bundle(for: FeedPresenter.self),
-           comment: "Title for the feed view"
-        )
-    }
-    
-    func didStartLoadingFeed() {
-        errorView.display(.noError)
-        loadingView.display(FeedLoadingViewModel(isLoading: true))
-
-    }
-    
-    func didFinishLoadingFeed(with feed:[FeedImage]) {
-        feedView.display(FeedViewModel(feed: feed))
-        loadingView.display(FeedLoadingViewModel(isLoading: false))
-    }
-    
-    func didFinishLoadingFeed(with error: Error){
-        errorView.display(.error(message: feedLoadError))
-        loadingView.display(FeedLoadingViewModel(isLoading: false))
-    }
-
-}
-
 class FeedPresenterTests: XCTestCase {
     
     func test_title_isLoacalized() {
